configure the temp directory to be used for streaming handoff
